# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GIQ HR Recruitment Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /interviews:
    get:
      operationId: listInterviews
      tags: [interviews]
      summary: List interviews
      parameters:
        - name: jobId
          in: query
          schema:
            type: string
        - name: candidateId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of interviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InterviewListResponse"
    post:
      operationId: scheduleInterview
      tags: [interviews]
      summary: Schedule an interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInterviewBody"
      responses:
        "201":
          description: Interview scheduled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
  /interviews/{id}:
    get:
      operationId: getInterview
      tags: [interviews]
      summary: Get interview by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    put:
      operationId: updateInterview
      tags: [interviews]
      summary: Update interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateInterviewBody"
      responses:
        "200":
          description: Updated interview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interview"
    delete:
      operationId: deleteInterview
      tags: [interviews]
      summary: Cancel interview
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /interviews/{id}/send-invite:
    post:
      operationId: sendInterviewInvite
      tags: [interviews]
      summary: Send interview invite email to candidate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # Analytics
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required:
        - total
        - page
        - limit
        - totalPages
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, recruiter]
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - email
        - name
        - role
        - isActive
        - createdAt
        - updatedAt
...
    Job:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        department:
          type: string
        location:
          type: string
        employmentType:
          type: string
          enum: [full_time, part_time, contract, internship]
        seniority:
          type: string
          enum: [entry, mid, senior, lead, executive]
        requiredSkills:
          type: array
          items:
            type: string
        preferredSkills:
          type: array
          items:
            type: string
        minExperience:
          type: integer
          nullable: true
        maxExperience:
          type: integer
          nullable: true
        minSalary:
          type: integer
          nullable: true
        maxSalary:
          type: integer
          nullable: true
        salaryCurrency:
          type: string
          nullable: true
        description:
          type: string
        responsibilities:
          type: string
        qualifications:
          type: string
        status:
          type: string
          enum: [draft, open, closed, archived]
        createdById:
          type: string
          nullable: true
        createdBy:
          $ref: "#/components/schemas/User"
          nullable: true
        candidateCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - department
        - location
        - employmentType
        - seniority
        - requiredSkills
        - preferredSkills
        - description
        - responsibilities
        - qualifications
        - status
        - candidateCount
        - createdAt
        - updatedAt
...
    Candidate:
      type: object
      properties:
        id:
          type: string
        fullName:
          type: string
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        experienceSummary:
          type: string
          nullable: true
        educationSummary:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, reviewing, shortlisted, interview_scheduled, rejected, hired]
        recruiterNotes:
          type: string
          nullable: true
        currentJobId:
          type: string
          nullable: true
        latestScore:
          type: number
          nullable: true
        latestFit:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - fullName
        - skills
        - status
        - createdAt
        - updatedAt
...
    Interview:
      type: object
      properties:
        id:
          type: string
        candidateId:
          type: string
        jobId:
          type: string
        candidate:
          $ref: "#/components/schemas/Candidate"
          nullable: true
        job:
          $ref: "#/components/schemas/Job"
          nullable: true
        interviewerName:
          type: string
        interviewType:
          type: string
          enum: [phone_screen, technical, behavioral, final, panel]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        location:
          type: string
          nullable: true
        meetingLink:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: string
          nullable: true
        inviteSentAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - candidateId
        - jobId
        - interviewerName
        - interviewType
        - scheduledAt
        - durationMinutes
        - status
        - createdAt
        - updatedAt
    InterviewListResponse:
      type: object
      properties:
        interviews:
          type: array
          items:
            $ref: "#/components/schemas/Interview"
        meta:
          $ref: "#/components/schemas/PaginationMeta"
      required:
        - interviews
        - meta
    ScheduleInterviewBody:
      type: object
      required: [candidateId, jobId, interviewerName, interviewType, scheduledAt, durationMinutes]
      properties:
        candidateId:
          type: string
        jobId:
          type: string
        interviewerName:
          type: string
        interviewType:
          type: string
          enum: [phone_screen, technical, behavioral, final, panel]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        location:
          type: string
        meetingLink:
          type: string
        notes:
          type: string
        sendInvite:
          type: boolean
          default: false
    UpdateInterviewBody:
      type: object
      properties:
        interviewerName:
          type: string
        interviewType:
          type: string
          enum: [phone_screen, technical, behavioral, final, panel]
        scheduledAt:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        location:
          type: string
        meetingLink:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled, no_show]
        notes:
          type: string
...